+++ packages/app/src/sections/leaderboard/AllTime.tsx
@@ -10,62 +10,13 @@ import Table, { TableHeader } from 'components/Table'
 import { TableCell } from 'components/Table/TableBodyRow'
 import { BANNER_HEIGHT_DESKTOP } from 'constants/announcement'
 import { DEFAULT_LEADERBOARD_ROWS } from 'constants/defaults'
-import useENSAvatar from 'hooks/useENSAvatar'
-import { StyledTrader } from 'sections/leaderboard/trader'
 import { selectShowBanner } from 'state/app/selectors'
 import { useAppSelector } from 'state/hooks'
 import { selectWallet } from 'state/wallet/selectors'
 import { FOOTER_HEIGHT } from 'styles/common'
 import media from 'styles/media'
-import { getMedal } from 'utils/competition'
-import { staticMainnetProvider } from 'utils/network'
 
-type TraderCellProps = {
-	onClickTrader: (trader: string) => void
-	compact?: boolean
-	rank: number
-	traderEns?: string | null
-	traderShort: string
-	trader: string
-}
-
-const TraderCell: FC<TraderCellProps> = ({
-	onClickTrader,
-	compact,
-	rank,
-	traderEns,
-	traderShort,
-	trader,
-}) => {
-	const avatar = useENSAvatar(staticMainnetProvider, traderEns)
-
-	return (
-		<StyledOrderType onClick={() => onClickTrader(trader)}>
-			{compact && rank + '. '}
-			<StyledTrader>
-				{avatar?.data ? (
-					<>
-						{!avatar.isLoading && avatar.data && (
-							<img
-								src={avatar.data}
-								alt={''}
-								width={16}
-								height={16}
-								style={{ borderRadius: '50%', marginRight: '8px' }}
-								// @ts-ignore
-								onError={(err) => (err.target.style.display = 'none')}
-							/>
-						)}
-						{traderEns}
-					</>
-				) : (
-					traderEns ?? traderShort
-				)}
-			</StyledTrader>
-			{getMedal(rank)}
-		</StyledOrderType>
-	)
-}
+import TraderCell from './TraderCell'
 
 type AllTimeProps = {
 	stats: AccountStat<Wei, number>[]
@@ -166,7 +117,6 @@ const AllTime: FC<AllTimeProps> = ({
 												onClickTrader={onClickTrader}
 												compact={compact}
 												rank={cellProps.row.original.rank}
-												traderEns={cellProps.row.original.traderEns}
 												traderShort={cellProps.row.original.traderShort}
 												trader={cellProps.row.original.trader}
 											/>
@@ -244,7 +194,6 @@ const AllTime: FC<AllTimeProps> = ({
 										onClickTrader={onClickTrader}
 										compact={compact}
 										rank={cellProps.row.original.rank}
-										traderEns={cellProps.row.original.traderEns}
 										traderShort={cellProps.row.original.traderShort}
 										trader={cellProps.row.original.trader}
 									/>
