+++ packages/app/src/state/exchange/actions.ts
@@ -1,4 +1,3 @@
-import { SynthExchange } from '@kwenta/sdk/types'
 import { toWei, truncateNumbers } from '@kwenta/sdk/utils'
 import { createAsyncThunk } from '@reduxjs/toolkit'
 import { wei } from '@synthetixio/wei'
@@ -466,13 +465,6 @@ export const setRatio =
 			})
 		}
 	}
-
-export const fetchWalletTrades = createAsyncThunk<SynthExchange[], void, ThunkConfig>(
-	'exchange/fetchWalletTrades',
-	(_, { extra: { sdk } }) => {
-		return sdk.exchange.getWalletTrades()
-	}
-)
 
 export const setExchangeRates =
 	(exchangeRates: Rates): AppThunk =>
